Add locale identification methods and relevant tests
In this commit, new functions have been added to identify the locale based on parameters such as 'accept-language', 'cookie', 'localStorage', and 'navigator'. This also includes a new feature in the getAcceptLanguage function to handle either a Headers object or a string to be more flexible. Test cases have been written for the new functionalities.